Port analyzer spec to Tape test
This is the last one!
